tutorials/talk-to-your-files-agent/requirements.txt
docling>=2.24.0
aixplain>=0.2.27
python-dotenv>=1.0.1
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.161.0
pytest>=8.3.4
tqdm>=4.67.1 